Move button and dropdown select to input dir
