+++ src/components/input/DropdownSelect.tsx
@@ -1,13 +1,15 @@
-import ArrowDown from '../assets/svg/arrow-down.svg?react';
+import ArrowDown from '../../assets/svg/arrow-down.svg?react';
 
 interface Props {
-  Name: React.FC;
+  Name: React.FC | string;
 }
 
 export default function DropdownSelect({ Name }: Props) {
+  const isText = typeof Name === 'string';
+
   return (
     <div className="flex items-center gap-1.5">
-      <Name />
+      {isText ? Name : <Name />}
       <ArrowDown />
     </div>
   );
